requirements.txt
pymupdf 
pdfplumber
pytesseract
pillow 
spacy
numpy
huggingface_hub




